Add frontend and backend code to handle stub test
