+++ src/pages.js
@@ -11,9 +11,10 @@ import { GuidedSourceDataPage } from "./stories/pages/guided-mode/data/GuidedSou
 import { GuidedMetadataPage } from "./stories/pages/guided-mode/data/GuidedMetadata"
 import { GuidedUploadPage } from "./stories/pages/guided-mode/options/GuidedUpload"
 import { GuidedDeleteFilesPage } from "./stories/pages/guided-mode/options/GuidedDelete"
-import { GuidedConversionTypePage } from "./stories/pages/guided-mode/options/GuidedConversionType"
+import { GuidedConversionOptionsPage, GuidedConversionTypePage } from "./stories/pages/guided-mode/options/GuidedConversionOptions"
 import { GuidedResultsPage } from "./stories/pages/guided-mode/results/GuidedResults"
 import { Dashboard } from "./stories/Dashboard"
+import { GuidedStubReviewPage } from "./stories/pages/guided-mode/options/GuidedStubReview"
 
 
 const overviewIcon = `
@@ -108,20 +109,20 @@ const pages = {
                 section: sections[1],
             }),
 
-            type: new GuidedConversionTypePage({
+            options: new GuidedConversionOptionsPage({
                 label: "Conversion type",
                 section: sections[2]
             }),
-            // upload: new GuidedUploadPage({
-            //     label: "Upload dataset",
-            //     section: sections[2]
-            // }),
+            review: new GuidedStubReviewPage({
+                label: "Upload dataset",
+                section: sections[2]
+            }),
             // deletefiles: new GuidedDeleteFilesPage({
             //     label: "Delete files",
             //     section: sections[2]
             // }),
 
-            review: new GuidedResultsPage({
+            approve: new GuidedResultsPage({
                 label: "Review results",
                 section: sections[3]
             }),
