+++ src/stories/pages/globals.js
@@ -1,1 +1,1 @@
-export const sections = ['Project Setup', 'Dataset Structure', 'Conversion Options', 'Results']
+export const sections = ['Project Setup', 'Dataset Structure', 'Configure Conversion', 'Results']
